% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPRS_functions.R
\name{create_n_2D_TPRS}
\alias{create_n_2D_TPRS}
\title{Create TPRS}
\usage{
create_n_2D_TPRS(size, spectral_density = 0.1, n)
}
\arguments{
\item{size}{The width or height of spatial surface. Size of TPRS becomes (size x size). Only makes square TPRS for now.}

\item{spectral_density}{Spectral density for Gaussian process that generates TPRS. Defaults to 0.1.}

\item{n}{Number of TPRS to generate to be added to regression.}
}
\value{
List with TPRS dataframe, omega 1 and 2 (Fourier frequency values corresponding the basis functions in expand.grid() format).
}
\description{
Create TPRS
}
\examples{
TPRS = create_n_2D_TPRS(16, .1, 10)
}
