% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPRS_functions.R
\name{simulate_gp}
\alias{simulate_gp}
\title{Simulate and (optionally) plot a Gaussian Process}
\usage{
simulate_gp(size, spectral_density, plot = TRUE)
}
\arguments{
\item{size}{Width/height of spatial grid}

\item{spectral_density}{Spectral density parameter for GP}

\item{plot}{Logical; whether to plot the simulated GP}
}
\value{
The simulated GP object
}
\description{
Simulate and (optionally) plot a Gaussian Process
}
\examples{
simulate_gp(128, 0.1, plot = TRUE)
}
