% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPRS_functions.R
\name{visualize_2D_TPRS}
\alias{visualize_2D_TPRS}
\title{Plot TPRS}
\usage{
visualize_2D_TPRS(tprs_obj)
}
\arguments{
\item{tprs_obj}{Object returned from create_n_2D_TPRS}
}
\value{
Nothing. Produces a plot of all TPRS basis components
}
\description{
Plot TPRS
}
\examples{
obj = create_n_2D_TPRS(16, .1, 10)
visualize_2D_TPRS(obj)
}
